moved part of code
